Refactor API routing structure
