+++ backend/app/routes/__init__.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+from fastapi import APIRouter
+
+from . import crops, price, recommend, refresh
+
+api_router = APIRouter()
+
+
+@api_router.get("/api/health")
+def health() -> dict[str, str]:
+    return {"status": "ok"}
+
+
+api_router.include_router(crops.router)
+api_router.include_router(recommend.router)
+api_router.include_router(price.router)
+api_router.include_router(refresh.router)
